Project metadata now has modules and summaries
Created the concept of modules, which are hierarchical groups of files in one or more directories, bundled together to help summarize the codebase. Modules have a min and max token count, to improve efficiency.

- vyb init now creates the initial set of modules and calls the LLM to create summaries for them;
- vyb update is a new command, that compares the persisted module with what's in the filesystem, and updates the annotations of the modules that have changed, and every module that depends on them;

Annotation generation is done in two phases:
- first we loop recursively from leafs to the root, updating the "self-contained" context, which is composed by:
-- Internal Context: a summary of the contents directly within the module;
-- Public Context: a summary of the exposed contents directly within the module, as well as any of its sub-modules;
- later, once all nodes have "self-contained" contexts, we call the LLM again, this time only with the annotations, so we can get the "External Context" of each module in the tree. This is the context of where in the hierarchy each module is located, and what it has access to;

This logic could be more efficient, and rely better on the LLM. For example, one possible improvement here would be for the request for summaries to provide the current summary of the module, and instruct the LLM to only make changes if strictly needed. Then we could use the change in the summary to determine whether we should update other summaries that depend on it or not.

The resulting summaries also are not distinct enough from each other and the prompts could use some tweaking.