feat: move serialization out of the payload package (#32)
Previously, the user message string was built in the payload package. This created a tight coupling between the calling code and the specific prompt format expected by the LLM.

With this change, the responsibility of serializing the request data into a user message is moved to the individual LLM provider implementations. This is achieved by introducing new request structs in the  package, which are then passed to the providers.

This refactoring improves the separation of concerns, making the code more modular and easier to maintain and extend.